Refactor registrar
