+++ uplink/clients/aiohttp_.py
@@ -13,7 +13,7 @@ from concurrent import futures
 import aiohttp
 
 # Local imports
-from uplink.clients import interfaces, registrar
+from uplink.clients import interfaces, register
 
 
 def threaded_callback(callback):
@@ -32,7 +32,7 @@ def threaded_callback(callback):
     return new_callback
 
 
-@registrar.handler
+@register.handler
 def _aiohttp_handler(client):
     if isinstance(client, aiohttp.ClientSession):
         return AiohttpClient(session=client)
